feat: Add Docker build and run scripts
This commit adds the Docker build and run scripts for the project. The `build.sh` script is used to build the Docker image with the tag `rolimoa:ubuntu22.04`, while the `run.sh` and `run_dev.sh` scripts are used to run the Docker container with various options and configurations. These scripts provide a convenient way to build and run the project in a Docker environment.