feat: Add Docker scripts for development and execution
This commit adds Docker scripts for building and running the project in a Docker environment. The `build.sh` script is used to build the Docker image with the tag `rolimoa:ubuntu22.04`, while the `run.sh` and `run_dev.sh` scripts are used to run the Docker container with various options and configurations. These scripts provide a convenient way to develop and execute the project using Docker.